add helper methods

+++ app/controllers/tweets_controller.rb
@@ -1,20 +1,16 @@
 class TweetsController < ApplicationController
 
+    skip_before_action :verify_authenticity_token
+  
     def index
       @tweets = Tweet.all
       render 'tweets/index'
     end
 
-    def create
-        token = cookies.signed[:twitter_user_session_token]
-        session = Session.find_by(token: token)
-    
-        if session
-          user = session.user
-          @tweet = user.tweets.new(tweet_params)
-          
+    def create   
+        if logged_in?
+          @tweet = current_user.tweets.new(tweet_params)
           if @tweet.save
-          
             render 'tweets/create'
           else
             render json: { success: false }
@@ -25,18 +21,25 @@ class TweetsController < ApplicationController
     end
 
     def destroy
-      @tweet = Tweet.find_by(id: params[:id])
       
-      if @tweet.destroy
-        render json: { success: true }
+      if logged_in?
+        @tweet = Tweet.find_by(id: params[:id])
+        if @tweet.destroy
+          render json: { success: true }
+        else
+          render json: { success: false }
+        end
       else
         render json: { success: false }
       end
+
     end
 
     private
 
     def tweet_params
-        params.require(:tweet).permit(:message, :user)
+        params.require(:tweet).permit(:message, :username, :id)
     end
+  
+
 end
